requirements.txt
flask==3.0.0
werkzeug==3.0.1
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.26.2
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==1.1.0
transformers==4.35.2
torch==2.1.1